zheniaBot/requirements.txt
aiogram==3.3.0
pika==1.3.2
protobuf==4.25.3
grpcio==1.62.0
grpcio-tools==1.62.0

